Add files via upload
This is the first draft of my consolidation project which shows the beginning of my approach to how I would complete my project. I started by creating a word bank for my project. Since I wanted to have a long list for my word bank, I decided to choose a theme for my word bank which was musical instruments. I chose this because I love to mess around with music and I believe that people would enjoy the category I chose for my game. After defining my word bank, I moved on to creating the functions needed for the program. I started by creating a main function to input the introduction for the players participating in the game. I then wrote in the function for displaying the words and having the program choose the word that the players have to guess at random. Afterward, I assigned the function "choose_word" to the operator "secret_word", made a list called "guessed_letters" to keep track of the guessed letters that each player guesses, and added an operator for the attempts each player makes which is 3 attempts. After this, I put in loops in my program so that my program can process the inputs of each player. I first ensured that the players' inputs had to be in lowercase since my words were in lowercase and included if statements for when players do not input an alphabetical letter or if they guess a letter they had already guessed. I then added an "append" function to include the previous that a player has attempted. After this, I added if statements based on whether a player guesses a letter correctly or incorrectly as well as printing how many attempts a player had. Afterward, I added a message that shows up when a player guesses the word correctly and then added a message if a player has run out of attempts and reveal the correct word.
After this, I checked to see if I had any problems with the code so I ran it through and found several errors and things I forgot to add. To fix this, I placed the display and random word choice functions before the main function which allowed for the program to run smoothly afterward. I realized that I had forgotten to put in the option to choose how many players are playing since the program starts up without asking how many players are playing. The program also included the position of the letters in the program which is what I was not supposed to do. In my next draft, I will make sure that I include the all the criteria in the fundamentals and start to include the optional parts into my program.